requirements.txt
yfinance>=0.2.18
ccxt>=4.2.0
backtrader>=1.9.78
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
alpha-vantage>=3.0.0
